+++ internal/service/eks/cluster.go
@@ -8,6 +8,7 @@ import (
 	"time"
 
 	"github.com/aws/aws-sdk-go/aws"
+	"github.com/aws/aws-sdk-go/service/ec2"
 	"github.com/aws/aws-sdk-go/service/eks"
 	"github.com/hashicorp/aws-sdk-go-base/v2/awsv1shim/v2/tfawserr"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/customdiff"
@@ -16,7 +17,7 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/validation"
 	"github.com/hashicorp/terraform-provider-aws/internal/conns"
 	"github.com/hashicorp/terraform-provider-aws/internal/flex"
-	"github.com/hashicorp/terraform-provider-aws/internal/service/ec2"
+	tfec2 "github.com/hashicorp/terraform-provider-aws/internal/service/ec2"
 	tftags "github.com/hashicorp/terraform-provider-aws/internal/tags"
 	"github.com/hashicorp/terraform-provider-aws/internal/tfresource"
 	"github.com/hashicorp/terraform-provider-aws/internal/verify"
@@ -157,6 +158,56 @@ func ResourceCluster() *schema.Resource {
 					},
 				},
 			},
+			"legacy_cluster_params": {
+				Type:     schema.TypeList,
+				Optional: true,
+				Computed: true,
+				MaxItems: 1,
+				Elem: &schema.Resource{
+					Schema: map[string]*schema.Schema{
+						"master_config": {
+							Type:     schema.TypeList,
+							Optional: true,
+							MaxItems: 1,
+							Elem: &schema.Resource{
+								Schema: map[string]*schema.Schema{
+									"high_availability": {
+										Type:     schema.TypeBool,
+										Required: true,
+										ForceNew: true,
+									},
+									"instance_type": {
+										Type:     schema.TypeString,
+										Required: true,
+										ForceNew: true,
+									},
+									"public_ip": {
+										Type:     schema.TypeString,
+										Optional: true,
+										ForceNew: true,
+									},
+									"volume_iops": {
+										Type:     schema.TypeInt,
+										Optional: true,
+										ForceNew: true,
+									},
+									"volume_size": {
+										Type:     schema.TypeInt,
+										Required: true,
+										ForceNew: true,
+									},
+									"volume_type": {
+										Type:         schema.TypeString,
+										Required:     true,
+										ForceNew:     true,
+										ValidateFunc: validation.StringInSlice(ec2.VolumeType_Values(), false),
+									},
+								},
+							},
+						},
+					},
+				},
+			},
 			"name": {
 				Type:         schema.TypeString,
 				Required:     true,
@@ -268,6 +319,10 @@ func resourceClusterCreate(d *schema.ResourceData, meta interface{}) error {
 		input.KubernetesNetworkConfig = expandNetworkConfigRequest(d.Get("kubernetes_network_config").([]interface{}))
 	}
 
+	if v, ok := d.GetOk("legacy_cluster_params"); ok {
+		input.LegacyClusterParams = expandLegacyClusterParams(v.([]interface{}))
+	}
+
 	if v, ok := d.GetOk("version"); ok {
 		input.Version = aws.String(v.(string))
 	}
@@ -376,6 +431,10 @@ func resourceClusterRead(d *schema.ResourceData, meta interface{}) error {
 		return fmt.Errorf("error setting kubernetes_network_config: %w", err)
 	}
 
+	if err := d.Set("legacy_cluster_params", flattenLegacyClusterParams(cluster.LegacyClusterParams)); err != nil {
+		return fmt.Errorf("error setting legacy_cluster_params: %w", err)
+	}
+
 	d.Set("name", cluster.Name)
 	d.Set("platform_version", cluster.PlatformVersion)
 	d.Set("role_arn", cluster.RoleArn)
@@ -388,7 +447,7 @@ func resourceClusterRead(d *schema.ResourceData, meta interface{}) error {
 
 	tags := KeyValueTags(cluster.Tags).IgnoreAWS().IgnoreConfig(ignoreTagsConfig)
 
-	//lintignore:AWSR002
+	// lintignore:AWSR002
 	if err := d.Set("tags", tags.RemoveDefaultConfig(defaultTagsConfig).Map()); err != nil {
 		return fmt.Errorf("error setting tags: %w", err)
 	}
@@ -501,7 +560,7 @@ func resourceClusterUpdate(d *schema.ResourceData, meta interface{}) error {
 		o, n := d.GetChange("tags_all")
 		// FIXME: Use eks.UpdateTags after TagResource and UntagResource are supported in C2 EKS API.
 		// To use EC2 API arn contains the cluster id.
-		if err := ec2.UpdateTags(connEc2, d.Get("arn").(string), o, n); err != nil {
+		if err := tfec2.UpdateTags(connEc2, d.Get("arn").(string), o, n); err != nil {
 			return fmt.Errorf("error updating tags: %w", err)
 		}
 	}
@@ -687,6 +746,64 @@ func expandLoggingTypes(vEnabledLogTypes *schema.Set) *eks.Logging {
 	}
 }
 
+func expandLegacyClusterParams(tfList []interface{}) *eks.LegacyClusterParams {
+	if len(tfList) == 0 {
+		return nil
+	}
+
+	tfMap, ok := tfList[0].(map[string]interface{})
+	if !ok || tfMap == nil {
+		return nil
+	}
+
+	legacyParams := &eks.LegacyClusterParams{}
+
+	if masterConfig, ok := tfMap["master_config"].([]interface{}); ok && len(masterConfig) > 0 {
+		legacyParams.MasterConfig = expandMasterConfig(masterConfig)
+	}
+
+	return legacyParams
+}
+
+func expandMasterConfig(tfList []interface{}) *eks.MasterConfig {
+	if len(tfList) == 0 {
+		return nil
+	}
+
+	tfMap, ok := tfList[0].(map[string]interface{})
+	if !ok || tfMap == nil {
+		return nil
+	}
+
+	masterConfig := &eks.MasterConfig{}
+
+	if v, ok := tfMap["high_availability"].(bool); ok {
+		masterConfig.HighAvailability = aws.Bool(v)
+	}
+
+	if v, ok := tfMap["instance_type"].(string); ok && v != "" {
+		masterConfig.MastersInstanceType = aws.String(v)
+	}
+
+	if v, ok := tfMap["public_ip"].(string); ok && v != "" {
+		masterConfig.MasterPublicIp = aws.String(v)
+	}
+
+	if v, ok := tfMap["volume_iops"].(int); ok && v != 0 {
+		masterConfig.MastersVolumeIops = aws.Int64(int64(v))
+	}
+
+	if v, ok := tfMap["volume_size"].(int); ok && v != 0 {
+		masterConfig.MastersVolumeSize = aws.Int64(int64(v))
+	}
+
+	if v, ok := tfMap["volume_type"].(string); ok && v != "" {
+		masterConfig.MastersVolumeType = aws.String(v)
+	}
+
+	return masterConfig
+}
+
 func flattenCertificate(certificate *eks.Certificate) []map[string]interface{} {
 	if certificate == nil {
 		return []map[string]interface{}{}
@@ -797,6 +914,35 @@ func flattenNetworkConfig(apiObject *eks.KubernetesNetworkConfigResponse) []inte
 	tfMap := map[string]interface{}{
 		"service_ipv4_cidr": aws.StringValue(apiObject.ServiceIpv4Cidr),
 		"ip_family":         aws.StringValue(apiObject.IpFamily),
+	}
+
+	return []interface{}{tfMap}
+}
+
+func flattenLegacyClusterParams(legacyParams *eks.LegacyClusterParams) []interface{} {
+	if legacyParams == nil {
+		return nil
+	}
+
+	tfMap := map[string]interface{}{
+		"master_config": flattenMasterConfig(legacyParams.MasterConfig),
+	}
+
+	return []interface{}{tfMap}
+}
+
+func flattenMasterConfig(masterConfig *eks.MasterConfig) []interface{} {
+	if masterConfig == nil {
+		return nil
+	}
+
+	tfMap := map[string]interface{}{
+		"high_availability": aws.BoolValue(masterConfig.HighAvailability),
+		"instance_type":     aws.StringValue(masterConfig.MastersInstanceType),
+		"public_ip":         aws.StringValue(masterConfig.MasterPublicIp),
+		"volume_iops":       aws.Int64Value(masterConfig.MastersVolumeIops),
+		"volume_size":       aws.Int64Value(masterConfig.MastersVolumeSize),
+		"volume_type":       aws.StringValue(masterConfig.MastersVolumeType),
 	}
 
 	return []interface{}{tfMap}